Introduce the "Use-case" top-level menu item with the generated list of solution pages
This commit also contains some cleanups for the layout for solution pages under
the new site. While not entirely finished this at least renders some of the
content we already show on the legacy version of the site within the context of
our new layouts

This doesn't yet clean up the "More" tab.