[INFRA-618] optimise the jpg
Run the image through photoshop to reduce the quality to 41.
This seems to be the best you can get before banding starts on the skies.

In reality this picture is probably not the best - as it is very fissy
with lots of harsh contrast changes and skies that makes it hard to
optimize.  Very little of it is also shown, so it is questionable if a
4k image is needed.  However I have not resized this (which would reduce
the file size drastically again as there may be a reason for the huge
dimensions that I am unaware of.