Introduce /project/ pages to give a more discoverable home to some of our projects
There are a couple initiatives within the Jenkins project which are are not
quite "core" or "plugins" in the traditional sense that we have had in the
project. These projects generally expand the horizons of Jenkins in a social or
technology aspect and as such I think require some escalated visibility inside
of the jenkins.io site structure.