Introduce /projects/ pages to give a more discoverable home to some of our projects (#260)
* Introduce /project/ pages to give a more discoverable home to some of our projects

There are a couple initiatives within the Jenkins project which are are not
quite "core" or "plugins" in the traditional sense that we have had in the
project. These projects generally expand the horizons of Jenkins in a social or
technology aspect and as such I think require some escalated visibility inside
of the jenkins.io site structure.

* Incorporate sub-projects into the top nav-bar

This required a little bit of shuffling around of content to make the bar look
less ridiculous. I'm still not thrilled with the "busy-ness" of it, but don't
have a better suggestion at this point in time