Add the [pipeline] block support for nicely enumerating either/or/both Declarative and Pipeline Script
This gives us a [pipeline] block to use inside of AsciiDoc files which will
automatically generate the appropriate output for HTML pages. This allows
writing documentation that references both Decalarative Pipeline and Pipeline
Script in the same blocks

For example:

    [pipeline]
    ----
    // Script //
    node {
        stage('Build') {
        checkout scm
        // Install dependencies
        sh 'npm install'
        }
    }

    // Declarative //
    pipeline {
        agent docker: 'node:6.3'
        stages {
            stage('Build') {
                sh 'npm install'
            }
        }
    }
    ----

The use of the `// Script //` and `// Declarative //` delimiters are required to
make the AsciiDoctor extension do its job properly. Without these special
delimiters the block will not render anything, as it cannot ascertain whether
the code included is Declarative Pipeline or Pipeline Script.