Update the handbook navigation to make the "next" link point to the next section or chapter
I find it a bit confusing to be on a chapter "index" page and have the
right-arrow navigation link take me to the next chapter. This change makes the
right-arrow link take the reader to the first section.

On the last section in the chapter, the right-arrow becomes the next chapter.

The left-arrow link will always be the previous section (if in the sections) or
the previous chapter (if in the chapter).

This should make it clearer how a reader should navigate through chapters and
sections by just clicking right-arrow, right-arrow, right-arrow