Fix outdated infomation on auto shared libraries
The automatic shared libraries section refers to a plugin that was
tombstoned and replaced by the main github-branch-source plugin.

So far as I have seen in a quick investigation the automatic loading
that the old Github origanization folder had was not part of the revamp
at the SCM API 2.0 changes.