Clean up how 'references' work in the changelog.
Previously, if both 'references' and top-level 'issue' and/or 'pull'
were set, both were displayed in independent sets of parentheses.

Now, if 'references' is defined, everything in it is shown and nothing
else. Only if 'references' doesn't exist, a top-level 'issue' is shown
and only if neither exists will a top-level 'pull' be shown.