Fix pagination for tags
The existing page linking logic only works for the general blogs list.
When looking at tags with multiple pages of results, the pagination
links back to the general blog pages instead of more tag results.

This change fixes this by using relative links to stay within the page
structure that it starts in.  It also adjusts the links generated depending
on if this the first page (which has a non-numbered url) or later pages
(which have numbers from 2 up.