Redirect to Wikipedia for Java 11 Web Start deprecation
Oracle announced in March 2018 that it will not include Java Web Start
in Java SE 11.  Jenkins provides a "Java Web Start" button to simplify
launching of clients with Java Web Start.  That simplified start technique
is often used for Windows agents.  Jenkins on Java 11 can't use that
button because the underlying Java implementation does not support it.

The communication protocol used by agent.jar continues to be fully
supported and operational with Java 11.  Only the Java Web Start button
from the web page is not supported in Java 11.

See [JENKINS-52282] and https://github.com/jenkinsci/jenkins/pull/3766
for more details.