Fix broken 'Connect' meetings links in SIG pages
The 'Connect' links are intended to link to locations outside the page.
They open a new browser tab when the link is clicked.

However, some of the special interest groups don't have a solo meeting
URL that is external to the page.  Rather than remove the 'Meetings'
link in the 'Connect' section for those pages, this change will resolve
those links to the 'Meetings' heading inside the page.

External meeting note locations:

* Advocacy and Outreach SIG links to external meeting notes
* Platform SIG links to external meeting notes

Internal meeting note locations:

* Chinese localization SIG
* Cloud native SIG
* Hardware / EDA SIG
* Pipeline authoring SIG