[WEBSTITE-610] Clarify credential masking limitations
Credential masking is a reasonable attempt to prevent accidental
disclosure of credentials.  It does not (and cannot) prevent those
with malicious intent from disclosing credentials which their Pipeline
is allowed to access.  If a Pipeline can use a credential, it has the
values of the credential and can leak those values if it chooses to do so.