Warn user against a local git repository
Blue Ocean attempts to support a local repository but will then have a
job defined which will fail on agents that cannot access the master
file system. Better for the reader to recommend a remote repository
rather than a local repository which will fail as they take later
recommended steps.