Add Platform Labeler as a featured plugin
Has 6 newbie-friendly issues that should be easy for a person that has
access to the type of Linux platform being sought.

I don't have a license for SUSE Linux Enterprise Server or for Red Hat
Enterprise Linux, but I suspect there are people in the community that
have access to those licenses and can extract the desired data from
them.