Add link to 'Contributing to Jenkins' slides
Contributing to a community is most likely to succeed when your
contribution aligns with one of your skills, interests, or abilities.

* If you’re interested, you’ll work better within the community
* If you’re using a skill that you’ve developed, you’ll be more
  involved in the community
* If you have a talent for something, you’ll be more likely to
  help the community with that talent

If you’re using a Jenkins component now, your help with that component
will increase your confidence in the component and will help users and
administrators in the community.