Blog post: Matrix building with scripted pipeline
This blog post covers how to perform matrix building with scripted
pipelines in Jenkins.  This is a supplemental followup post for
declarative matrix pipelines which explains to users of scripted
pipelines how to also perform matrix builds.

Additionally, it covers exposing matrix building to users as a shared
library step.