Rework parameters fragment file for presentation
Narrow the example strings so they do not push the table of contents
off the right edge of the page.

Prefer sentence per line for easier and smaller editing later.

Split the parameters into networking, miscellaneous, and Java properties.

Remove some incorrect text and text that is no longer relevant.

Could not remove the reference to the system properties table.  It will
need a separate pull request to move it from the wiki to jenkins.io.