fix #2950 - Python tutorial fails in pyinstaller
Requires the addition of a step that creates a new directory on each
new build so that pyinstaller has a clean directory to prepare the
distribution.  That step needs a copy of the compiled code, so a `stash`
step is added to the compile phase.

Includes an environment block to reduce the length of the lines in
the tutorial.  Explanation of the environment block is included.

Includes a description of BUILD_ID and a link to environment variables.

Includes a cleanup step in the post section to remove the artifacts
generated by the 'Deliver' stage.  Without that step, each job leaks
more files than necessary.

If there is concern at the addition of an unexplained cleanup step,
it could be removed without harming the tutorial.