Link in network discovery services
I reviewed https://github.com/jenkinsci/jenkins/pull/4460 and found no
links that were not already included in the network discovery text.

I thought a link to the 2.204.2 upgrade guide might be helpful to some
readers, but it is misleading because it indicates that there is an
'escape hatch', while this change intentionally removes the 'escape
hatch'.