JENKINS-56643 changelog entry is a developer change
The original motivation for PR 3941 was to assure that the symbolic
links used for lastSuccessfulBuild and lastFailedBuild would not have
brief periods when they were not visible on the file system due to a
remove and recreate step.

That original motivation was no longer needed for lastFailedBuild or
lastSuccessfulBuild because PR-3982 removed the use of symbolic links
for those files.  That change has been included in Jenkins since 2.185.

PR-3941 adjusts the Util.createSymlink API call.  That API is no
longer used for lastSuccessfulBuild and lastFailedBuild.