Easier site dev and test without unnecessary https://jenkins.io/
When a hyperlink in a file refers directly to https://jenkins.io/, it
makes testing and development more difficult.  When the author clicks
one of those absolute links during development on their own computer,
they will navigate to the page on jenkins.io and no longer be navigating
their development copy.

This removes several of the absolute references from documents that are
more frequently edited.  It does not remove the absolute references from
blog posts or from security advisories.  Those are less frequently edited
and less likely to confuse authors during development.