Fix many broken hyperlinks on jenkins.io (#3140)
* Use Jetty current for JMX and ALPN docs

Rather than trying to track the exact Jetty version that we're using
and the exact Jetty version that is still documented on the Jetty
site, let's use 'current'.  Less likely to be broken in the future.

* Fix broken link to GSoC gitter channel

* Use github.com rather than github.org in URL link

* Fix link to Chinese site

* Remove evergreen - broken link, halted project

* Fix broken OWASP links

* Fix link to Jenkins X documentation

* Use valid HTTP URL for meetings, not invalid HTTPS

The http://meetings.jenkins-ci.org/ site does not provide HTTPS.
Using an HTTPS URL does not help the reader of the material. If they
click the link, they want to see the page, not the warning that the
site does not serve pages with HTTPS.

* Fix broken link to Assembla webhook docs

* Fix GSoC report blog broken link

* Fix links to Hacktoberfest

* Fix link to credentials storage blog post

* Link to wiki exporter rather than word 'here'

* Link to Kafka remoting plugin instead of TODO plugin