Link to Google Doc in the 'Connect' section
The Pipeline Authoring SIG provides meeting notes in a Google Doc, using
the same pattern as several other Jenkins Special Interest Groups.
The "Connect" section of each SIG page provides links to resources
that are outside the jenkins.io site.  Other SIG's place a link to
their meeting notes in the "Connect" section and rely on the generated
table of contents to link to the description of meeting times and other
information related to the meeting.