Clarify plugin manager upgrade guide statements
Upgrade guide is not used to describe new features.  It is
intentionally limited to areas in the upgrade which may imapact users
and administrators.

Describing the intentional difference in the "Available" tab of the
plugin manager may help some administrators understand that they need
to enter a value before they will see available plugins.