Remove disclaimers for ancient Jenkins versions
No need to warn readers about issues with Jenkins versions that are more
than a year old.  Jenkins long term support (LTS) releases are delivered
every three months.  When a new LTS is delivered, the Jenkins project
expects users to upgrade to that new LTS.  We continue to provide update
information for releases that are as much as 12 months old.  Releases that
are more than 12 months old may be dropped from the Jenkins update center.

Jenkins 2.80 is almost three years old.  Jenkins 2.54 is even older
than that.  Readers are expected to run a recent LTS release or a recent
weekly release.