Avoid WEBSITE-630 non-deletion of removed files
Replacement content of case study with a page noting that the case
study has been removed.

The URL to the case study will still include the name of the company that
provided the case study, but there will be no mention of the company name
in the list of blog posts.  The prior author reference to an employee
of the company has been removed as well so that the page will be less
likely to be associated with the company.

Once we resolve WEBSITE-630, this page can be deleted.