Move 3 topics into the "Scaling" section
The "Architecting for Manageability", "Architecting for Scale", and
"Hardware Recommendations" were not directly available from the User
Handbook links on the left of the page.  This pull request moves those
pages under the "Scaling Jenkins" heading and provides redirects from
their previous locations.

The pull request also converts "master" to "controller" in the modified
files.  References to the "master key" remain, because the "master key"
is stored on the controller but it is the key for the entire Jenkins
installation, not just for the controller.

There are still many things that can be improved in this documentation.
They can be handled in future pull requests once this structure and
terminology update is merged.