Various changes for formatting
Use YAML source instead of bash source for YAML files so that the
Asciidoc highlighting will help the reader detect important content in
the YAML file.

Use text source for files that are not shell scripts and not yaml.
Hope to avoid formatting content that does not have a specific syntax.

Mark the text blocks with the name of the file they are describing so
that the reader is shown a visual cue of the name of the file.

Fix a YAML syntax error in a label.

Narrow the text examples so that they do not push the table of contents
off screen on 1280 pixel wide web browsers.