Add anchor redirects inside installing index page
We prefer to use the page level redirect mechanism described in the
contributing guide.  However, this page needs to remain as a page yet
have internal anchors within the page redirected to other locations.
The anchor redirects are targeting several different pages.  This
Javascript technique allows entrance into a single page to redirect to
a different page based on the anchor text.

The Javascript that is embedded in this page is taken from
https://stackoverflow.com/questions/1305211/javascript-to-redirect-from-anchor-to-a-separate-page/21198129#21198129
with deepest gratitude for that answer.

I've checked each of the redirects using a local web page that opened
http://localhost:4242/doc/book/installing/#windows (and others like it).
They are well-behaved running my local development copy of jenkins.io.
When this is merged, I will perform additional checks to confirm that
the technique also works for https://www.jenkins.io/doc/book/installing/