Remove default values for GitSCM class
`doGenerateSubmoduleConfigurations: false` is the default.
`doGenerateSubmoduleConfigurations: true` is untested and unlikely to work.

We hope in a future release of the git plugin to stop suggesting it as
an option in the pipeline syntax generator without breaking backwards
compatibility.

submoduleCfg: [] is the default.
It is only honored if `doGenerateSubmoduleConfigurations: true` is set.