Fix redirect to "Jenkins says me reverse proxy setup is broken"
The redirect from inside Jenkins core uses
https://www.jenkins.io/redirect/troubleshooting/broken-reverse-proxy to
provide diagnostic hints to the user.  Without this fix, that redirect
is broken.

This change also splits the reverse proxy configuration page into two
pages, one with the general introduction and navigation links and the
other with the troubleshooting section. It includes additional redirects
in case there are pages that link from internal page references to the
original page.

Thanks to @DaGeRe for reporting the issue