Add Jenkins is the Way video to root page
The Jenkins Governance Board meeting Oct 21, 2020 approved that the video
may be placed on the root page of https://jenkins.io below the Jumbotron.

The video is intentionally placed below the the features segment and above
the events calendar.  It is "edge to edge" to match with the Jumbotron
layout and provide visual separation for the calendar and blog content
that follows the video.

Meeting notes at:

https://docs.google.com/document/d/11Nr8QpqYgBiZjORplL_3Zkwys2qK1vEvK-NYyYa4rzg/edit#

Video recording of meeting at:

https://youtu.be/zBOwRubg7FE?list=PLN7ajX_VdyaMsOs5pRWGByvIcuJOpHpC6&t=1034