Better phrasing of JDK change and fix a typo
The phrasing for the JDK change from OpenJDK to AdoptOpenJDK was not
clear enough.  We've already switched from OpenJDK to AdoptOpenJDK for
JDK 11.  This change is making JDK 8 consistent with JDK 11.

Does not mention the additional benefit from updating package versions
in the operating system.  For example, there are valuable performance
improvements in the transition from command line git 2.11 (Debian 9)
to command line git 2.20 (Debian 10).