Use update and upgrade more consistently in this section
In this section, the word "update" describes the replacement of an
older Jenkins plugin version with a newer Jenkins plugin version.
That usage is specific to this section, not a general practice that
anyone should assume is required in other sections of the
documentation.

In this section, the word "upgrade" describes the replacement of a
Jenkins core version with a newer version of Jenkins core.  That usage
is specific to this section, not a general practice.