[JENKINS-65069] Refer to AdoptOpenJDK CA issue for Debian 10 Docker image
https://issues.jenkins.io/browse/JENKINS-65069 reports that the user was
unable to register a custom certificate authority in the Docker image
beginning with Jenkins 2.263.4.

The issue is known to AdoptOpenJDK and a fix is in progress.
See https://github.com/AdoptOpenJDK/openjdk-installer/issues/105 for
the latest status.

A workaround is provided in issue 8243 of the jitsi-meet project.
See https://github.com/jitsi/jitsi-meet/issues/8243#issuecomment-744181944

Special thanks to Oleg Nenashev for locating the work around and to
Hashim for reporting the issue and investigating it so effectively.