Fix #4167 Redirect Algolia search id into same page
The Javascript that redirects references from inside one page (like
installing#windows) to another page (like installing/windows) will
redirect the user to a page "undefined" (ike installing/undefined).
This is an experiment that hopes to redirect the search result to the
same page.

Works in my local test environment, but won't be 100% confident until
I've seen it work on the jenkins.io site.