Include office hours and meetings headings
They are separate concepts and should remain separate in the document.
We are reusing office hours as the location for the SIG meeting
because it is more effective to do that.  Don't want to remove the
heading because the concepts of office hours and SIG meetings need to
remain, even if they happen in the same meeting currently.