Use correct method in example @DataBoundSetter (#4318)
If a plugin is wanting an optional field to be null if it's not set, it
will need to use Util.fixEmpty (which will coerce a null or empty
string to null) instead of Util.fixNull (which will coerce a null string
to an empty string). Fix the example so it works properly.