Include docker build steps in Windows install
Feedback from a reader noted that step 4 in the Windows instructions for
Docker was relying on the reader to page backwards in the document to find
step 4 in the Linux/macOS instructions.  That's unclear to readers and
makes the experience difficult.

Better to include the content directly in the text so that the user
consistently scrolls through the instructions rather than needing to
backtrack into a section for a platform that may not be interesting
to them.

Same change has been made in the Docker tutorials.  I confirmed that the
instructions work on my Windows 10 computer running the latest Docker
Desktop for Windows.