Remove OpenJ9 JVM support from roadmap (#4466)
The hotspot JVM is available for JDK 11 on all the platforms currently
of interest to the Jenkins Platform SIG.  It is available for:

* AMD64/Intel64
* Arm64
* PowerPC64
* s390x

The prototype OpenJ9 image generation has been removed from the Jenkins
Docker images.  Work is now proceeding to create multi-architecture
Docker images using `docker bake`.

See https://github.com/jenkinsci/docker/pull/1134 for details.

No further progress is expected on support of the OpenJ9 virtual machine
in the Jenkins project.  Let's remove the entry from the roadmap.