Use blue ocean 1.25.0 in sample Docker images
Blue Ocean 1.25.0 was released 3 days ago with a fix for a downstream
build link.  I've been running it successfully on my instance for the
several days without any issue.  Let's use the latest release so that
tutorial users do not have so many warnings for plugin upgrades.

https://github.com/jenkinsci/blueocean-plugin/releases/tag/blueocean-parent-1.25.0
provides more details of the content of the release.

https://issues.jenkins.io/browse/JENKINS-60995 is the issue that was
fixed by the release.