Adapt Linux install docs to current status
The fontconfig package is already a dependency of the operating system
Java installer.  It is only needed when using a JDK that is not provided
as an operating system packages (like Eclipse Adoptium).

LTS 2.319.3 requires the `daemonize` package on Red Hat and its
derivatives, but is provided by EPEL.  EPEL instructions are different
depending on the operating system version.  Stop trying to describe EPEL
installation, point the reader to the EPEL installation instructions.