Add block to deny access to files in root
The previous config allowed anyone access to files in the Jenkins root dir, many of them containing secret tokens & other sensitive info. I think that the official docs should provide good defaults.

I'm not an nginx expert, and while this has worked fine in my environment, this config may break some plugins that rely on serving files from the root dir over HTTP, but I thought it better to provide a fairly restrictive default rather than rely on a blacklist of files (since the directory is a bit of a Wild West, with different Jenkins core features and plugins storing files in it).