JDK distro is Eclipse Temurin, distributed by Eclipse Adoptium (#4989)
https://adoptium.net/faq.html#temurinName says that:

What is this "Eclipse Temurin" name?

This is the project and brand name for the binaries that the Eclipse
Foundation produces. You can think of these as the successor binaries
to AdoptOpenJDK. However, it is important to note that these binaries
are just one of many that will eventually be in the Eclipse Adoptium
Marketplace. While we appreciate that the Adoptium/Temurin name split is
more confusing that just “Adoptium”, this is similar to how other
vendors brand their binaries, e.g. Amazon has Corretto, Azul has Zulu
(and others). The “Adoptium” project and working group will deal
with more than just Temurin so the distinction is important to maintain.