Use a 3 part Blue ocean admonition
Show selected portions of the Blue Ocean admonition based on the
context.  If the page refers to Blue Ocean in general terms, use only
the first paragraph of the admonition.  If the page refers to Blue Ocean
Pipeline visualization, then include the portion of the admonition for
Pipeline visualation.  If the page refers to Blue Ocean Pipeline
creation, then include the portion of the admonition for Pipeline
creation.

Thanks to the Docs office hours (Asia) for the suggestion to make the
admonitiion content based on context.

Thanks to Gavin Mogan for the request to make it clear that Blue Ocean
is not being actively developed.

Thanks to Kevin Martens for the original Blue Ocean admonition pull
request.