Correct English language and clarify some phrases (#5345)
I chose the following (on line 439):
The naming of a solution page template
but I also considered:
The naming of the solution page template
I think the first one is the correct one, but I am not really sure.

I wonder whether "C++" is being processed correctly. If does not appear to be visible in the Github preview when a double quotes follows C++.

What is the style for lower-case? Both lower-case and lowercase are in this document. I changed lowercase to lower-case on line 486 because there are more "lower-case" with hyphens than without in this document, but personally I think lowercase is better.

I deleted "For meetup" on line 526. Perhaps, it was the start of a sentence, but there was no clue what that might be.

Line 595 is not clear ("into" maybe should be "in" or "to"), but I am not sure how to correct it, so I did not change anything there.