Correct the Apache reverse proxy configuration with mod_rewrite (#5358)
* Correct the Apache reverse proxy configuration with mod_rewrite

The reverse proxy check in Jenkins uses URL-encoded slashes, which Apache URL-encodes by default. This change adds the "NoEscape" flag on the rewrite rule to prevent that. It also adds a small reminder of the HTTPS options.

* Fix space URLs handling in the mod_rewrite Apache setup

When using the NE flag, Apache will not encode spaces properly. We enforce its encoding with the B flag. The BNP flag makes sure we use %20 instead of + to encode the space (Jetty rejects URLs with a + in it, but accepts those with %20).