Remove java.level property from examples
The `java.level` property is deprecated and will be removed from a
future plugin parent pom.

See https://github.com/jenkinsci/plugin-pom/pull/522 for the plugin pom
deprecation of `java.level`.

See the following additional references:

* https://github.com/jenkinsci/plugin-pom/pull/478#discussion_r809561384
* https://github.com/jenkinsci/plugin-pom/pull/478#issuecomment-1082144103
* https://github.com/jenkinsci/plugin-pom/pull/522#issuecomment-1083320869

Also uses newer Jenkins version in the example.  Would have liked to use
the automatic replacement as is done in the "Choosing a Jenkins version"
baseline page, but that was a larger change than I'm ready to make in
this pull request.