Add information about silent install on Windows (#5478)
* Add information about silent install

This adds information on performing a silent install using the MSI installer on Windows. It gives an overview of the properties that can be used to override default values during the installation and some examples of how to use them on the command line.

* Fix list 

Fixes the list and rewords one area of the instructions.