Update administering-jenkins-on-kubernetes.adoc
I have made this document taking reference and guidance from @MarkEWaite
Please let me know if this document is viable or it needs any changes.
The https://www.jenkins.io/doc/book/system-administration/administering-jenkins-on-kubernetes/ contains nothing and is empty. So we can put this information in that section for the time being.
If any changes are needed in this section, we can later update it by raising a issue. This is because, an issue suggesting that this section is empty already exists, so by putting this we can close that issue.