Update reverse-proxy-configuration-with-jenkins.adoc (#5509)
* Update reverse-proxy-configuration-with-jenkins.adoc

Just a small change which was mentioned in the Issue #5491 
To completely close this issue, We will have to change the layouts of the files in this section. I am not sure how to exactly do it.
But if someone guides me on the following on how I can change the layout and which files I need to update, I'll do it ASAP, or someone else can help in changing the layout too. Please refer to the Issue #5491 for more details.

* Update content/doc/book/system-administration/reverse-proxy-configuration-with-jenkins.adoc

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>