Simplify testing with site local links
When the https://www.jenkins.io/ URL is embedded in the source code of
the jenkins.io pages, it causes a user exploring a pull request through
the preview sites to be moved from the preview site to the
www.jenkins.io site.  If they are not very carefully watching the URL
address bar, they can waste time looking at the production site when
they wanted to review the preview site.

Text replacement was done with `sed -i` for files that contained the
literal text `link:https://www.jenkins.io/ and were inside the content
directory and were not in the content/security subdirectories.

Two preceding commits were made to correct other errors detected during
the review of this change.

The other changes were:

    Fix broken links to changelog

    Changelog links expire after a period of time and are moved into
    changelog-old.  Need to update links elsewhere when that move happens

Another change was:

    Fix broken slideshare link