Add berviantoleo as a mentor (#5847)
Hi,

I'm interested in the project "Docker-based Jenkins quickstart examples"
.

I would like to add myself as a mentor for that project. I am also open
to having a co-mentor.

I made the `docker-compose.yml` 2 months ago in
[here](https://gist.github.com/berviantoleo/b149ff1902c7b07668c5c7d8af455c6d).
It's just an example, and we will need to add more things to this
project because the project size might be no more than 175 hours.