Remove version specifier from demo plugin installs (#5865)
The version specifiers like `blueocean:1.26.0` were originally added
because it is a strongly preferred practice to precisely define the
versions of plugins that are being installed.  However, the default
behavior of the plugin installation manager is to install the most
recent release of the plugin, even if a plugin version is specified.

If a user wants to install an older version, the `--latest false`
argument is used.  However, that installs outdated versions of many
plugins when the plugins and their versions are not fully specified.

Using `blueocean` is better than `blueocean:1.26.0` in this case because
it avoid a precise version specification that is then ignored by the
plugin installation manager and instead relies on the plugin
installation manager to choose the most recent release that is supported
on this Jenkins version.