Downgrade git and git client plugins in tutorials
https://github.com/jenkins-infra/jenkins.io/issues/5864 correctly
reports that git client plugin 4.0.0 provides a newer JGit version than
the blue ocean 1.26.0 plugin is ready to use for Pipeline creation with
git.

This is a short term workaround to allow users to continue running the
tutorials while we work on the better solution.

Pipeline creation from the Blue Ocean interface is a rarely used
feature in most Jenkins installations.  We can make it work in this
demonstration configuration for now and then remove this workaround in a
future change.