Rewrite PHP guide to use Jenkinsfile based setup (#5885)
The current guide uses an ant based setup and does not include the
actual ant build file. The original site it was sourced from has been
archived. The ant file in the (now archived) repo for the source site is
massive and almost certainly out of date. I think that these factors
make the current guide hard to follow for those just starting out.

This rewrite is based on the setup I am currently actively using.

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>