doc: `user-site` no longer needed (#5893)
## Description

When a pull request is made to the jenkins.io repository, a full preview
of the site is now available for anyone to view. This preview accurately
reflects the final layout of the site, unlike GitHub pages which may not
be an exact match. Developers no longer need to maintain separate GitHub
pages for this purpose.

Signed-off-by: Rajiv Ranjan Singh <rajivperfect007@gmail.com>
Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>