Allow up to 2 elected board members from a single company (#5689)
## Allow up to 2 elected board members from a single company

Governance board needs members that are actively involved in the Jenkins
project. By specifically allowing up to 2 elected board members from a
single company, we do not create a majority of elected members of the
board from a single company.

The issue comes because the board has declared previously that Kohsuke
is affiliated with CloudBees. His permanent membership on the board and
his affiliation with CloudBees means that only one other person from
CloudBees can serve on the board. The board needs more active
participants to assist with the work of the board. Allowing 1 more board
member from CloudBees or any other company that already has a member on
the board will strengthen the board and its work to serve the Jenkins
project.

Needs approval of the governance board before this pull request is
merged.