Only check dependencies once a month
The plugin archetype now includes dependabot checks once a month.
That is sufficient for a plugin that is maintained less frequently and
avoids the distraction of dependabot pull requests every week.

Also removes the "insert maintainers here" because there have been
several cases where a new contributor did not replace that text with
the name of a maintainer.