Prevent absolute links from adoc to www.jenkins.io and jenkins.io (#5899)
## Prevent absolute links to jenkins.io and www.jenkins.io in adoc files

When the https://www.jenkins.io/ URL is embedded in the source code of
the jenkins.io pages, it causes a user exploring a pull request through
the preview sites to be moved from the preview site to the
www.jenkins.io site. If they are not very carefully watching the URL
address bar, they can waste time looking at the production site when
they wanted to review the preview site.

* https://github.com/jenkins-infra/jenkins.io/issues/5718
* https://github.com/jenkins-infra/jenkins.io/pull/5715

@ADITYADAS1999 I assumed that the issue was more difficult than you were
ready to attack. If you'd like to review the technique I've used, I'd
love to have comments or recommendations of ways it could be improved.

Co-authored-by: Damien Duportal <damien.duportal@gmail.com>