Promote GitLab plugin modernization project to draft state (#5968)
Cleans up this project proposal so that it can be ready for draft state,
addressing previous concerns. Reduced the scope of the project to be
consistent with a 175-hour project size and clarified the difficulty
level: the project is "easy" in one sense but long and tedious in
another. Clarified the expected outcome, documented known challenges,
and provided links to other examples of this migration.

CC @jmMeessen