Add placeholder to indicate additional information is needed (#6036)
We've seen a few issues reported that lack any kind of information about
the identity of the affected plugin. Our guess is that they were
submitted through this form, with submitters not realizing that, unlike
in JENKINS, there's no component identifying the plugin.

This adds a fixed placeholder to the summary as a first step, let's see
how well this works.