suggestions for FOSDEM blog
These are some suggestions I had for updating the blog post, adding in some links to make it more interactive, and formatting the quotes a bit more. I also suggested resizing some of the images so that they are more aligned, and added alt text for the images to provide for accessibility. I'm not 100% sure that the alt text is correct, so if there is something better that fits, I'm all for it!

@gounthar I noticed the photo used in your entry is of the person soldering a chip on your agent, would you want to update the alt text for that if that is not the case though?

Ultimately, if the changes are disliked, please feel free to revert the commit. I tried to retain all of the quotes as they had been written previously.