Update index.html.haml
I am writing to propose a fix for broken links in the  repository. While reviewing the repository, I noticed that the links on lines 133 and 199 of a particular file were not working. I have since made the necessary changes to update these links and ensure that they are functional.

To make these changes, I forked the  repository and cloned it to my local machine. I then made the necessary updates to the file, testing the changes to ensure that the links were working properly. I have committed the changes to my local Git repository and pushed the changes to my forked repository on GitHub.

I am now creating a pull request to propose these changes to the repository.