Remove details of irrelevant CVE
Continue to include the link, but follow the pattern used in other
locations of just including the link rather than including link and
explanation text.

The Spring Security update includes CVE links but does not waste the
space to describe issues that do not affect Jenkins.  Use that same
pattern.