suggestions for text update
Hi @gounthar, this is a revised/suggested version of the post. I did my best to keep your voice/message as present as possible, and hope that nothing is too egregious.

I left the list of SSH options as is for the most part, but since several of them are either older/not recommended, is there a reason to keep them there beyond listing _any_ option? The SSHelper one for instance, hasn't been updated since 2020 and you explicitly say that you do not recommend it. Is there another version/option that we could list instead, or would it be possible to remove anything that we wouldn't want users doing?