suggestions for syntax/content of blog
Hi @gounthar, I was able to review the post and everything looks great! I made some suggestions/edits to try and clean up a few things/add a link or two, but this is fantastic. I also updated the filename to be aligned with the current date. Would it be possible to update the actual folder in this case as well?