Jenkins.io - add image guidelines to contributing.adoc (#6239)
This pull request is to add some image guidelines for contributing to
Jenkins. These are some general ideas meant to help clarify how
screenshots/images can be used and added to the documentation. It also
provides guidelines on adding alt text, to improve the SEO score and
provide further accessibility for users.

I have also updated a fair amount of the sentence/page formatting to use
single sentence per line (in a lot of areas). I know this is not
complete throughout the documentation, but plan on making further edits
when adding additional guideline suggestions (documentation formatting
and book submissions).